2_Quick_Admin_Suite/CDK_DataCollection/requirements.txt
aws-cdk-lib==2.44.0
constructs>=10.0.0